Rename the lib directories from converse to dialogue
