+++ lib/dialogue.rb
@@ -1,18 +1,18 @@
-require "converse/conversation_factory"
-require "converse/conversation_options"
-require "converse/conversation_template"
-require "converse/conversation_template_runner"
-require "converse/conversation"
-require "converse/conversation_handler"
-require "converse/conversation_options_validator"
-require "converse/invalid_options_error"
-require "converse/dsl"
-require "converse/message_decorators"
-require "converse/storable"
-require "converse/streams"
-require "converse/template_already_registered_error"
-require "converse/template_factory"
-require "converse/version"
+require "dialogue/conversation_factory"
+require "dialogue/conversation_options"
+require "dialogue/conversation_template"
+require "dialogue/conversation_template_runner"
+require "dialogue/conversation"
+require "dialogue/conversation_handler"
+require "dialogue/conversation_options_validator"
+require "dialogue/invalid_options_error"
+require "dialogue/dsl"
+require "dialogue/message_decorators"
+require "dialogue/storable"
+require "dialogue/streams"
+require "dialogue/template_already_registered_error"
+require "dialogue/template_factory"
+require "dialogue/version"
 
 module Converse
   class << self
