+++ lib/dialogue/conversation_factory.rb
