+++ lib/dialogue/conversation_handler.rb
