+++ lib/dialogue/conversation_options.rb
