+++ lib/dialogue/conversation_options_validator.rb
