+++ lib/dialogue/conversation_template.rb
