+++ lib/dialogue/conversation_template_runner.rb
