+++ lib/dialogue/dsl.rb
