+++ lib/dialogue/invalid_options_error.rb
