+++ lib/dialogue/message_decorators.rb
