+++ lib/dialogue/message_decorators/slack.rb
