+++ lib/dialogue/storable.rb
