+++ lib/dialogue/streams.rb
