+++ lib/dialogue/streams/slack.rb
