+++ lib/dialogue/template_already_registered_error.rb
