+++ lib/dialogue/template_factory.rb
