+++ lib/dialogue/version.rb
