+++ spec/spec_helper.rb
@@ -1,7 +1,7 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 
 require "rspec"
-require "converse"
+require "dialogue"
 
 Dir["./spec/support/**/*.rb"].sort.each { |f| require f }
